requirements.txt
transformers
torch
fastapi
uvicorn
faiss-cpu
numpy
pandas
scikit-learn